Change API of module to avoid left open files
Instead of accepting filenames, all types have constructors that accept
IOStreams instead. This should encourage users to use do-blocks to
ensure the file is properly closed, or at least it absolves this module
of that responsibility.

Also making the ContinuousFile type not exported, and creating array
constructors that accept IOStreams directly. This should make use of the
module more straight forward.